requirements.txt
pdfplumber>=0.10.0
tabula-py>=2.8.0
pandas>=2.0.0
camelot-py[cv]>=0.11.0
PyPDF2>=3.0.0
python-dateutil>=2.8.0
openpyxl>=3.1.0
